+++ hydromt/data_catalog/sources/dataset.py
@@ -85,7 +85,7 @@ class DatasetSource(DataSource):
             return None
 
         if isinstance(self.uri_resolver, ConventionResolver):
-            self._check_multifile_is_complete(uris, ds, handle_nodata)
+            self.uri_resolver._check_multifile_is_complete(uris, ds, handle_nodata)
 
         return self.data_adapter.transform(
             ds,
