+++ hydromt/data_catalog/sources/geodataset.py
@@ -83,7 +83,7 @@ class GeoDatasetSource(DataSource):
         if ds is None:  # handle_nodata == ignore
             return None
         if isinstance(self.uri_resolver, ConventionResolver):
-            self._check_multifile_is_complete(uris, ds, handle_nodata)
+            self.uri_resolver._check_multifile_is_complete(uris, ds, handle_nodata)
 
         return self.data_adapter.transform(
             ds,
